17.03 seo optimization

+++ crawl4me/src/main/java/by/njjnex/logic/FileUtils.java
@@ -7,11 +7,13 @@ public class FileUtils {
 	public static void deleteDir(String path) {
 		File dir = new File(path);
 		File[] filelist = dir.listFiles();
-		for (File file : filelist) {
-			if (file.isFile()) {
-				file.delete();
-			} else {
-				deleteDir(file.toString());
+		if (filelist != null) {
+			for (File file : filelist) {
+				if (file.isFile()) {
+					file.delete();
+				} else {
+					deleteDir(file.toString());
+				}
 			}
 		}
 		dir.delete();
